+++ serve/src/index.js
@@ -4,6 +4,7 @@ let app = new PIXI.Application({
 document.body.appendChild(app.view)
 
 let control = new PIXI3D.CameraOrbitControl(app.view)
+control.enableDamping = true
 
 app.loader.add("assets/chromatic/diffuse.cubemap")
 app.loader.add("assets/chromatic/specular.cubemap")